--- .idea/libraries/Maven__com_github_milkbowl_VaultAPI_master_8dc0859cba_1.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.github.milkbowl:VaultAPI:master-8dc0859cba-1">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/github/milkbowl/VaultAPI/master-8dc0859cba-1/VaultAPI-master-8dc0859cba-1.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/github/milkbowl/VaultAPI/master-8dc0859cba-1/VaultAPI-master-8dc0859cba-1-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/github/milkbowl/VaultAPI/master-8dc0859cba-1/VaultAPI-master-8dc0859cba-1-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
