--- .idea/libraries/Maven__javax_persistence_persistence_api_1_0.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: javax.persistence:persistence-api:1.0">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/javax/persistence/persistence-api/1.0/persistence-api-1.0.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/javax/persistence/persistence-api/1.0/persistence-api-1.0-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/javax/persistence/persistence-api/1.0/persistence-api-1.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
