+++ ButtonCore/src/main/java/buttondevteam/component/commands/CommandComponent.java
@@ -1,0 +1,16 @@
+package buttondevteam.component.commands;
+
+import buttondevteam.lib.architecture.Component;
+import org.bukkit.plugin.java.JavaPlugin;
+
+public class CommandComponent extends Component { //TODO: Do we just move everything here?
+	@Override
+	public void register(JavaPlugin plugin) {
+
+	}
+
+	@Override
+	public void unregister(JavaPlugin plugin) {
+
+	}
+}
